Add most_similar.py.
This script, given a binary and a set of other binaries for comparison,
will use difflib to narrow down the search (using quick_ratio) and then
identify the closest matching files from the filtered list.

Optionally, both the filtering and ranking can be done based on the
absolute number of matching bytes (i.e., scaling the ratio based on the
lengths of the pair of files). Alternatively, the length of the longest
contiguous match can be used for ranking.